Move ChainType types to Network
